+++ amberdata/src/endpoint/price.ts
@@ -1,0 +1,33 @@
+import { Requester, Validator } from '@chainlink/external-adapter'
+import { ExecuteWithConfig, Config } from '@chainlink/types'
+
+export const Name = 'price'
+
+const customError = (data: any) => {
+  return Object.keys(data.payload).length === 0
+}
+
+const customParams = {
+  base: ['base', 'from', 'coin'],
+  quote: ['quote', 'to', 'market'],
+}
+
+export const execute: ExecuteWithConfig<Config> = async (input, config) => {
+  const validator = new Validator(input, customParams)
+  if (validator.error) throw validator.error
+  const jobRunID = validator.validated.id
+
+  const coin = validator.validated.data.base
+  const market = validator.validated.data.quote
+  const url = `/api/v2/market/spot/prices/pairs/${coin.toLowerCase()}_${market.toLowerCase()}/latest`
+
+  const params = {
+    includeCrossRates: true,
+  }
+
+  const reqConfig = { ...config.api, params, url }
+
+  const response = await Requester.request(reqConfig, customError)
+  response.data.result = Requester.validateResultNumber(response.data, ['payload', 'price'])
+  return Requester.success(jobRunID, response)
+}