+++ blockcypher/src/adapter.ts
@@ -16,11 +16,9 @@ export const execute: ExecuteWithConfig<Config> = async (request, config) => {
   const jobRunID = validator.validated.id
   const endpoint = validator.validated.data.endpoint || DEFAULT_ENDPOINT
 
-  let result
   switch (endpoint) {
     case balance.Name: {
-      result = await balance.execute(config, request)
-      break
+      return balance.makeExecute(config)(request)
     }
     default: {
       throw new AdapterError({
@@ -30,12 +28,6 @@ export const execute: ExecuteWithConfig<Config> = async (request, config) => {
       })
     }
   }
-
-  return Requester.success(jobRunID, {
-    data: { result },
-    result,
-    status: 200,
-  })
 }
 
 export const makeExecute: ExecuteFactory<Config> = (config) => {