+++ cryptoapis/test/price.test.ts
@@ -1,9 +1,12 @@
-const { assert } = require('chai')
-const { assertSuccess, assertError } = require('@chainlink/adapter-test-helpers')
-const { execute } = require('../adapter')
+import { assert } from 'chai'
+import { Requester, AdapterError } from '@chainlink/external-adapter'
+import { assertSuccess, assertError } from '@chainlink/adapter-test-helpers'
+import { AdapterRequest } from '@chainlink/types'
+import { makeExecute } from '../src/adapter'
 
-describe('execute', () => {
+describe('price endpoint', () => {
   const jobID = '1'
+  const execute = makeExecute()
 
   context('successful calls @integration', () => {
     const requests = [
@@ -33,13 +36,11 @@ describe('execute', () => {
     ]
 
     requests.forEach((req) => {
-      it(`${req.name}`, (done) => {
-        execute(req.testData, (statusCode, data) => {
-          assertSuccess({ expected: 200, actual: statusCode }, data, jobID)
-          assert.isAbove(data.result, 0)
-          assert.isAbove(data.data.result, 0)
-          done()
-        })
+      it(`${req.name}`, async () => {
+        const data = await execute(req.testData as AdapterRequest)
+        assertSuccess({ expected: 200, actual: data.statusCode }, data, jobID)
+        assert.isAbove(data.result, 0)
+        assert.isAbove(data.data.result, 0)
       })
     })
   })
@@ -59,11 +60,13 @@ describe('execute', () => {
     ]
 
     requests.forEach((req) => {
-      it(`${req.name}`, (done) => {
-        execute(req.testData, (statusCode, data) => {
-          assertError({ expected: 400, actual: statusCode }, data, jobID)
-          done()
-        })
+      it(`${req.name}`, async () => {
+        try {
+          await execute(req.testData as AdapterRequest)
+        } catch (error) {
+          const errorResp = Requester.errored(jobID, new AdapterError(error))
+          assertError({ expected: 400, actual: errorResp.statusCode }, errorResp, jobID)
+        }
       })
     })
   })
@@ -81,11 +84,13 @@ describe('execute', () => {
     ]
 
     requests.forEach((req) => {
-      it(`${req.name}`, (done) => {
-        execute(req.testData, (statusCode, data) => {
-          assertError({ expected: 500, actual: statusCode }, data, jobID)
-          done()
-        })
+      it(`${req.name}`, async () => {
+        try {
+          await execute(req.testData as AdapterRequest)
+        } catch (error) {
+          const errorResp = Requester.errored(jobID, new AdapterError(error))
+          assertError({ expected: 500, actual: errorResp.statusCode }, errorResp, jobID)
+        }
       })
     })
   })